backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyPDF2
python-docx
httpx
